Images are now in a common folder.
Images are now in 'common'. Instead of embedding the image within the rkt file, i use 'bitmap/file' to load the img into some constant. this is the better way to do it so that the image can be shared amongst files and you dont see a bunch of image junk in .rkt files.